backend/requirements.txt
# ── Project Veracity — Backend Dependencies ───────────────────────────────────
# Install with: pip install -r requirements.txt

# Core API Framework
fastapi
uvicorn[standard]

# Data Validation
pydantic
python-dotenv

# AI / ML
transformers
torch
sentence-transformers

# Vector Database
chromadb

# HTTP Client (async LLM calls)
httpx

# ── Dev / Testing ──────────────────────────────────────────────────────────────
pytest
pytest-asyncio
pypdf
